+++ src/main/java/com/example/toyproject/TestDataInit.java
@@ -33,9 +33,10 @@ public class TestDataInit {
 
     @PostConstruct
     public void init() {
+        //Member member = Member.builder().email("test@gmail.com").password("12345678").name("test").build();
         SignUpDto signUpDto = new SignUpDto("test@gmail.com", "12345678", "test");
         memberService.signUp(signUpDto);
-
+        //아래 줄부터 error 발생
         List<Member> members = memberRepository.findByEmailAndPassword("test@gmail.com", "12345678");
         Member member = members.stream().findFirst().get();
         MemberInfoDto memberInfoDto = MemberInfoDto.builder()
@@ -54,7 +55,7 @@ public class TestDataInit {
     private List<BoardDto> createBookList() {
         List<BoardDto> list = new ArrayList<>();
         list.add(BoardDto.builder()
-                .title("testdatainit title1")
+                .title("title1")
                 .content("content1 test test test")
                 .rate(3)
                 .createdDate(LocalDateTime.now())
@@ -62,7 +63,7 @@ public class TestDataInit {
                 .isbn("1234-1234-1234-1234")
                 .dtype("B").build());
         list.add(BoardDto.builder()
-                .title("testdatainit title2")
+                .title("title2")
                 .content("content2 test test test")
                 .rate(3)
                 .createdDate(LocalDateTime.now())
@@ -70,7 +71,7 @@ public class TestDataInit {
                 .isbn("1234-1234-1234-1234")
                 .dtype("B").build());
         list.add(BoardDto.builder()
-                .title("testdatainit title3")
+                .title("title3")
                 .content("content3 test test test")
                 .rate(3)
                 .createdDate(LocalDateTime.now())
