+++ src/main/java/com/example/toyproject/service/BoardServiceImplV2.java
@@ -45,6 +45,7 @@ public class BoardServiceImplV2 implements BoardService {
         boardDto.setMember(memberByDto);
         Board board = boardDto.toEntity();
         Board savedBoard = boardRepository.save(board);
+        log.info("board saved!");
         return savedBoard.getId();
     }
