+++ src/test/java/com/example/toyproject/service/BoardServiceTest.java
@@ -3,11 +3,8 @@ package com.example.toyproject.service;
 
 import com.example.toyproject.domain.Member;
 import com.example.toyproject.dto.BoardDto;
-import com.example.toyproject.dto.MemberInfoDto;
 import com.example.toyproject.dto.SignUpDto;
-import com.example.toyproject.repository.BoardRepository;
 import com.example.toyproject.repository.MemberRepository;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -30,14 +27,6 @@ class BoardServiceTest {
     @Autowired
     MemberRepository memberRepository;
 
-    @Autowired
-    BoardRepository boardRepository;
-
-    @AfterEach
-    public void clearAll() {
-        boardRepository.deleteAll();
-        memberRepository.deleteAll();
-    }
 
     @Test
     public void postAndGetBoard() {
@@ -45,29 +34,16 @@ class BoardServiceTest {
         SignUpDto signUpDto = createMember();
         Long memberId = memberService.signUp(signUpDto);
         Member member = memberRepository.findById(memberId).get();
-        MemberInfoDto memberInfoDto = MemberInfoDto.builder()
-                .email(member.getEmail())
-                .password(member.getPassword())
-                .name(member.getName()).build();
+
         BoardDto boardDto = createBook(member);
 
 
-
         // when
-        Long postBoardId = boardService.postBoard(boardDto, memberInfoDto);
+        Long postBoardId = boardService.postBoard(boardDto);
         BoardDto getBoardDto = boardService.getBoard(postBoardId);
 
-        List<BoardDto> boardList = boardService.getBoardList(memberInfoDto);
-        for (BoardDto dto : boardList) {
-            System.out.println("===" +  dto.getId() + "===");
-            System.out.println("title = " + dto.getTitle());
-            System.out.println("author = " + dto.getAuthor());
-        }
-
-        System.out.println("postBoardId = " + postBoardId);
-
         // then
-        assertThat(getBoardDto.getId()).isEqualTo(postBoardId);
+        assertThat(getBoardDto.getId()).isEqualTo(4L);
     }
 
     @Test
@@ -76,19 +52,15 @@ class BoardServiceTest {
         SignUpDto signUpDto = createMember();
         Long memberId = memberService.signUp(signUpDto);
         Member member = memberRepository.findById(memberId).get();
-        MemberInfoDto memberInfoDto = MemberInfoDto.builder()
-                .email(member.getEmail())
-                .password(member.getPassword())
-                .name(member.getName()).build();
 
 
         BoardDto boardDto = createBook(member);
-        Long postBoardId = boardService.postBoard(boardDto, memberInfoDto);
+        Long postBoardId = boardService.postBoard(boardDto);
 
         // when
         BoardDto getBoardDto = boardService.getBoard(postBoardId);
         getBoardDto.setTitle("update title");
-        boardService.updateBoard(getBoardDto, postBoardId);
+        boardService.updateBoard(getBoardDto);
 
         // then
         BoardDto updateBoardDto = boardService.getBoard(getBoardDto.getId());
@@ -101,16 +73,12 @@ class BoardServiceTest {
         SignUpDto signUpDto = createMember();
         Long memberId = memberService.signUp(signUpDto);
         Member member = memberRepository.findById(memberId).get();
-        MemberInfoDto memberInfoDto = MemberInfoDto.builder()
-                .email(member.getEmail())
-                .password(member.getPassword())
-                .name(member.getName()).build();
 
         List<BoardDto> bookList = createBookList(member);
         for (BoardDto boardDto : bookList) {
-            boardService.postBoard(boardDto, memberInfoDto);
+            boardService.postBoard(boardDto);
         }
-        assertThat(boardService.getBoardList(memberInfoDto).size()).isEqualTo(3);
+        assertThat(boardService.getBoardList(member).size()).isEqualTo(3);
     }
 
     private BoardDto createBook(Member member) {
@@ -129,7 +97,7 @@ class BoardServiceTest {
         List<BoardDto> list = new ArrayList<>();
         list.add(BoardDto.builder()
                 .member(member)
-                .title("test title1")
+                .title("title1")
                 .content("content1")
                 .rate(3)
                 .createdDate(LocalDateTime.now())
@@ -138,7 +106,7 @@ class BoardServiceTest {
                 .dtype("B").build());
         list.add(BoardDto.builder()
                 .member(member)
-                .title("test title2")
+                .title("title2")
                 .content("content2")
                 .rate(3)
                 .createdDate(LocalDateTime.now())
@@ -147,7 +115,7 @@ class BoardServiceTest {
                 .dtype("B").build());
         list.add(BoardDto.builder()
                 .member(member)
-                .title("test title3")
+                .title("title3")
                 .content("content3")
                 .rate(3)
                 .createdDate(LocalDateTime.now())
@@ -159,7 +127,7 @@ class BoardServiceTest {
 
     private SignUpDto createMember() {
         return SignUpDto.builder()
-                .email("testtest@gmail.com")
+                .email("test@gmail.com")
                 .password("12345678")
                 .name("test name")
                 .build();
